24.05.17(금) Ajax방식파일업로드

+++ spring_boot_work/ex06/src/main/java/com/docmall/demo/util/FileUtils.java
@@ -13,12 +13,15 @@ import javax.imageio.ImageIO;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Component;
 import org.springframework.util.FileCopyUtils;
 import org.springframework.web.multipart.MultipartFile;
 
 import net.coobird.thumbnailator.Thumbnails;
 
-//여러 Controller에서 사용될수 있는 기능이므로 재사용성을 위해 클래스로 빼버렸음.
+//여러 Controller에서 사용될수 있는 기능이므로 클래스로 빼버렸음.(재사용성)
+
+@Component  // 스프링에서 클래스를 자동관리. bean으로 관리 할 수 있게 해준다. 
 public class FileUtils {
 
 	// [기능1]현재폴더를 운영체제에 맞게 문자열로 반환.
@@ -61,13 +64,17 @@ public class FileUtils {
 		File file = new File(uploadFoledr, dateFolder);  //(예) ":\\Dev\\upload\\pds"(고정)   "2024\\05\\16"(동적)
 		
 		// "2024/05/16" 폴더가 존재하지 않으면, 폴더 생성
+		// 새로운 날짜에 첫 파일업로드가 진행되면, 폴더생성되고, 두번째 파일업로드부터는 폴더가 생성되지 않든다.
 		if(file.exists() == false) {
-			file.mkdirs();  //폴더가 빠져도 다 생성하겠다.
+			file.mkdirs();  //폴더가 빠져도 다 생성하겠다. 
+			 //mkdirs는 폴더를 여러개 생성
+			 // mkdir은 폴덜르 1개만 생성(개발에서 잘 쓰지 않는다.)
 		}
 		
 		//클라이언트에서 보낸 파일명    UUID는 고유식별자를 말하며 고유한 값을 생성해줌.
 		String clientFileName = uploadFile.getOriginalFilename(); // abc.png		
 		UUID uuid = UUID.randomUUID(); //2f48f241-9d64-4d16-bf56-70b9d4e0e79a(인터넷에서 가져온 아무 고유값)
+									   //랜덤으로 값을 생성해준다. 중복이 나올수는 있으나 거의 나오지 않는다.
 		
 		//2f48f241-9d64-4d16-bf56-70b9d4e0e79a_abc.png
 		realUploadFileName = uuid.toString() + "_" + clientFileName;
@@ -81,16 +88,20 @@ public class FileUtils {
 			//원본파일 해상도크기를 줄여 섬네일이미지 생성하기
 			if(checkImageType(saveFile)) {
 				// Thumnail 파일명 : 2f48f241-9d64-4d16-bf56-70b9d4e0e79a_abc.png 생성
-				File thumnailFile = new File(saveFile, "s_" + realUploadFileName);
+				// thumnailFile 객체 : "C:\\Dev\\upload\\pds" "2f48f241-9d64-4d16-bf56-70b9d4e0e79a_abc.png"
+				File thumbnailFile = new File(file, "s_" + realUploadFileName);
 			
+				//saveFile객체 : 업로드 된 파일정보
 				BufferedImage bo_img = ImageIO.read(saveFile);
+				
+				
 				double ratio = 3;
 				int width = (int) (bo_img.getWidth() / ratio);
 				int height = (int) (bo_img.getHeight() / ratio);
 				
-				Thumbnails.of(saveFile)
+				Thumbnails.of(saveFile) 
 						  .size(width, height)
-						  .toFile(thumnailFile);
+						  .toFile(thumbnailFile);  
 			}
 		}catch(Exception e) {
 			e.printStackTrace();
@@ -100,7 +111,7 @@ public class FileUtils {
 	
 
 	// [기능3] 업로드파일의 MIME타입 확인. 즉 이미지 파일 또는 일반파일 여부를 체크
-	private static boolean checkImageType(File saveFile) {
+	public static boolean checkImageType(File saveFile) {
 
 		boolean isImageType = false;
 		try {
@@ -119,16 +130,18 @@ public class FileUtils {
 	
 	
 	// [기능4]이미지파일을 웹브라우저 화면에 보이는 작업.
-	// <img src="abc.gif"> <img src="매핑주소">
+	// <img src="abc.gif"> <img src="매핑주소"> 매핑주소를 통한 서버측에서 받아오는 바이트배열을 이용하여 브라우저가 이미지를 표시한다.
 	/*
 	  - String uploadPath : 서버 업로드폴더  예)"C:\\Dev\\upload\\pds"
 	  - String fileName : 이미지 파일명 (날짜폴더명 포함)
 	 */
+	
+	// 파일업로드되는 폴더가 외부 프로젝틍 존재하여, 보안적인 이슈가 있으므로, 업로드 파일들을 byte배열로 읽어서 클라이언트로 보낸다.
 	public static ResponseEntity<byte[]> getFile(String uploadPath, String fileName) throws Exception{
 		ResponseEntity<byte[]> entity = null;
 		File file = new File(uploadPath, fileName);  // 이 경로는 실제 파일이 존재 해야 함.
 		
-		if(!file.exists()) {
+		if(!file.exists()) { //파일이 존재하지 않으면 entity를 반환한다.
 			return entity;
 		}
 		
@@ -147,19 +160,20 @@ public class FileUtils {
 	/*
 	 - String uploadPath : 서버업로드 폴더
 	 - String folderName : 날짜 폴더명
-	 - String fileName : 파일명
+	 - String fileName : 파일명(날자폴더명 포함)
 	*/
 	
-	public static void delete(String uploadPath, String folderName ,String fileName) {
+	public static void delete(String uploadPath, String fileName, String type) {
 		
 		// 1)원본파일  예)"C:/Dev/upload/pds"  "2024/05/06"     "2f48f241-9d64-4d16-bf56-70b9d4e0e79a_abc.png"
-		File file1 = new File((uploadPath + folderName + "\\" + fileName).replace('\\', File.separatorChar));
-		if(file1.exists()) file1.delete();
+		File file1 = new File((uploadPath + "\\" + fileName).replace('\\', File.separatorChar));
+		if(file1.exists()) file1.delete();  //파일이 존재하면 삭제한다.
 
+		if(type.equals("image")) {
 		// 2) thumnail파일
-		File file2 = new File((uploadPath + folderName + "\\" + "_s" + fileName).replace('\\', File.separatorChar));
-		if(file2.exists()) file2.delete();
-
+			File file2 = new File((uploadPath + "\\" + "_s" + fileName).replace('\\', File.separatorChar));
+			if(file2.exists()) file2.delete();  // 썸네일파일이 존재하면 삭제한다.
+		}
 		
 	}
